feat: BREAKING CHANGES!!! change solution structure
Up to this commit, plugins and libraries have been attached as external folders to this project. This was convenient for debugging, but required having all projects on disk. Now the main solution does not contain them. For plugin development and debugging you need to create a separate project and include Asv.Drones project as a submodule. The standard plugin name should start with Asv.Drones.Sdr.Plugin.<NAME>.dll. You will also need to override the Asv.Drones.Sdr.Custom.props file, which should be located in the plugin solution folder.