feat: add OSD telemetry rate and display updates
Implemented an On-Screen Display (OSD) telemetry rate parameter and updated the way telemetry data is presented in the OSD for 'GPWorkMode', 'LlzWorkMode', and 'VorWorkMode' of the project 'Asv.Drones.Sdr'. The telemetry rate is now adjustable and the display information includes the total DDM, SDM, and Azimuth depending on the mode. The changes allow users to customize the frequency of telemetry data updates and improve the visibility of important values.

Asana: https://app.asana.com/0/1203851531040615/1206083632668454/f