feat: add comments for clarity and maintainability
Added necessary comments to classes and methods to describe their purpose and functionality in many files of the Asv.Drones.Sdr.Core and related projects. These comments contribute to the clarity and maintainability of the code, assisting both current and future developers in understanding the codebase logic and structure.

Asana: https://app.asana.com/0/1203851531040615/1206239158556872/f